backend-20: * moved mail transport creation to separate file
